refactor: move authToken app context to SplashActivity.kt
